DAOS-000 server: change nr_xs_helpers to be total number of helper XS
Change nr_xs_helpers from number of helper XS per target to be total
number of helper XS per server.
That allows multiple targets share the helper XS.
For example with config of "targets: 8" and "nr_xs_helpers: 4" then the
4 helper XS will be shared by the 8 targets.

Only for test, some details need to be refined if need to land it.

Signed-off-by: Xuezhao Liu <xuezhao.liu@intel.com>